# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List client appointments
      description: Returns the appointments booked by the signed-in client.
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      description: Books a new consultation appointment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      description: Updates an existing appointment (e.g. status change or reschedule).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /lawyer/clients:
    get:
      operationId: listLawyerClients
      tags: [lawyer]
      summary: List lawyer clients
      description: Returns the clients of the signed-in lawyer.
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createLawyerClient
      tags: [lawyer]
      summary: Add a client
      description: Creates a new client for the signed-in lawyer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid client payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lawyer/appointments:
    get:
      operationId: listLawyerAppointments
      tags: [lawyer]
      summary: List lawyer appointments
      description: Returns the appointments of the signed-in lawyer.
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LawyerAppointment"
    post:
      operationId: createLawyerAppointment
      tags: [lawyer]
      summary: Create a lawyer appointment
      description: Books a new appointment for one of the lawyer's clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LawyerAppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LawyerAppointment"
        "400":
          description: Invalid appointment payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lawyer/appointments/{id}:
    patch:
      operationId: updateLawyerAppointment
      tags: [lawyer]
      summary: Update a lawyer appointment
      description: Updates the status or payment state of one of the lawyer's appointments.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LawyerAppointmentUpdate"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LawyerAppointment"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    AppointmentStatus:
      type: string
      enum: [upcoming, completed, cancelled, pending, rescheduled]
    Appointment:
      type: object
      properties:
        id:
          type: string
        lawyerName:
          type: string
        lawyerAvatar:
          type: string
        lawyerGradient:
          type: string
        specialization:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: number
        mode:
          type: string
          enum: [online, offline]
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        caseType:
          type: string
        notes:
          type: string
        fee:
          type: number
        meetLink:
          type: string
      required:
        - id
        - lawyerName
        - lawyerAvatar
        - lawyerGradient
        - specialization
        - date
        - time
        - duration
        - mode
        - status
        - caseType
        - notes
        - fee
    AppointmentInput:
      type: object
      properties:
        lawyerName:
          type: string
        lawyerAvatar:
          type: string
        lawyerGradient:
          type: string
        specialization:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: number
        mode:
          type: string
          enum: [online, offline]
        caseType:
          type: string
        notes:
          type: string
        fee:
          type: number
        meetLink:
          type: string
      required:
        - lawyerName
        - lawyerAvatar
        - lawyerGradient
        - specialization
        - date
        - time
        - duration
        - mode
        - caseType
        - notes
        - fee
    AppointmentUpdate:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        duration:
          type: number
        mode:
          type: string
          enum: [online, offline]
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        caseType:
          type: string
        notes:
          type: string
        meetLink:
          type: string
...
    ClientStatus:
      type: string
      enum: [active, closed, pending]
    ClientVisibility:
      type: string
      enum: [public, lawyers_only, private]
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        initials:
          type: string
        gradient:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        caseType:
          type: string
        status:
          $ref: "#/components/schemas/ClientStatus"
        visibility:
          $ref: "#/components/schemas/ClientVisibility"
        lastContact:
          type: string
        totalSessions:
          type: number
        amountPaid:
          type: number
        nextAppointment:
          type: string
        notes:
          type: string
      required:
        - id
        - name
        - initials
        - gradient
        - email
        - phone
        - city
        - caseType
        - status
        - lastContact
        - totalSessions
        - amountPaid
        - notes
...
    ClientInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        caseType:
          type: string
        status:
          $ref: "#/components/schemas/ClientStatus"
        lastContact:
          type: string
        totalSessions:
          type: number
        amountPaid:
          type: number
        notes:
          type: string
      required:
        - name
        - email
        - phone
        - city
        - caseType
        - status
        - notes
    LawyerAppointmentStatus:
      type: string
      enum: [upcoming, completed, cancelled, pending]
    LawyerAppointment:
      type: object
      properties:
        id:
          type: string
        clientName:
          type: string
        clientInitials:
          type: string
        clientGradient:
          type: string
        caseType:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: number
        mode:
          type: string
          enum: [online, offline]
        status:
          $ref: "#/components/schemas/LawyerAppointmentStatus"
        fee:
          type: number
        isPaid:
          type: boolean
        notes:
          type: string
        meetLink:
          type: string
      required:
        - id
        - clientName
        - clientInitials
        - clientGradient
        - caseType
        - date
        - time
        - duration
        - mode
        - status
        - fee
        - isPaid
        - notes
    LawyerAppointmentUpdate:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        duration:
          type: number
        mode:
          type: string
          enum: [online, offline]
        status:
          $ref: "#/components/schemas/LawyerAppointmentStatus"
        isPaid:
          type: boolean
        notes:
          type: string
        meetLink:
          type: string
    LawyerAppointmentInput:
      type: object
      properties:
        clientName:
          type: string
        clientInitials:
          type: string
        clientGradient:
          type: string
        caseType:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: number
        mode:
          type: string
          enum: [online, offline]
        status:
          $ref: "#/components/schemas/LawyerAppointmentStatus"
        fee:
          type: number
        isPaid:
          type: boolean
        notes:
          type: string
        meetLink:
          type: string
      required:
        - clientName
        - clientInitials
        - clientGradient
        - caseType
        - date
        - time
        - duration
        - mode
        - status
        - fee
        - isPaid
        - notes
...